done with filters

+++ src/main/java/com/vonnueAmazonClone/amazonClone/DTO/ProductDto.java
@@ -38,5 +38,5 @@ public class ProductDto {
 
     private Seller seller;
 
-    private double averageRating;
+    private int averageRating;
 }
